% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross.R
\name{Cross}
\alias{Cross}
\title{Cross table}
\usage{
Cross()
}
\value{

}
\description{
Cross table
}
\examples{
Cross()
Cross()  |> groutable::dt2()
library(ggplot2)
ggplot(data= data.frame(x=1:12,y=1:12), aes(x,y)) +
  geom_point(shape = Cross()$symbols, size = 8) +
  theme_void()
}
\author{
Zhonghui Gai
}
