% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Greak.R
\name{Greek_alphabet}
\alias{Greek_alphabet}
\title{Greek alphabet table}
\usage{
Greek_alphabet()
}
\value{

}
\description{
Greek alphabet table
}
\examples{
Greek_alphabet()  |> groutable::dt2()
ggplot(data= data.frame(x=1:24,y=1:24), aes(x,y)) +
geom_point(shape = Greek_alphabet()$upper, size = 8) +
    theme_void()
ggplot(data= data.frame(x=1:24,y=1:24), aes(x,y)) +
    geom_point(shape = Greek_alphabet()$lower, size = 8) +
    theme_bw()+
    xlab("IL1\u03b2") +
    theme(axis.title.x = element_text(face = "bold.italic"),
          axis.title = element_text(size = 20)) +
   ylab("δ✠φυαμΣ")
}
\author{
Zhonghui Gai
}
