\name{nnumattr}
\Rdversion{1.1}
\alias{nnumattr}
\title{Number of numeric variables}
\description{
Counts number of numeric variables in data frame.
}
\usage{
nnumattr(df)
}
\arguments{
  \item{df}{data frame.}
}
\details{
Counts number of numeric variables in data frame.
}
\value{
Returns number of numeric variables.
}
\author{Jaroslav Myslivec <jaroslav.myslivec@upce.cz>}
