\name{symbols-package}
\Rdversion{1.1}
\alias{symbols-package}
\alias{symbols}
\docType{package}
\title{Symbol plots}
\description{Package that implements various symbol plots (bars, profiles, stars, Chernoff faces, color icons, stick figures).}
\details{
\tabular{ll}{
Package: \tab symbols\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2009-11-28\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{Jaroslav Myslivec <jaroslav.myslivec@upce.cz>}
\keyword{ package }
