% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_unicode.R
\name{test_unicode}
\alias{test_unicode}
\title{Unicode symbols in R}
\usage{
test_unicode(start = "25a0", end = "25ff", ...)
}
\arguments{
\item{start}{the start}

\item{end}{the end}

\item{...}{}
}
\value{

}
\description{
Unicode symbols in R
}
\examples{
test_unicode()
test_unicode(9500,9900)
plot(0, 0, main=bquote(symbol("\326")))
plot(0, 0, main=bquote("\u03b1"))
}
\author{
Zhonghui Gai
}
